+++ packages/blocks/core/js/wordpress/quote/index.js
@@ -30,9 +30,27 @@ export const Quote: BlockType = {
 				force: true,
 			},
 		},
-		'core/paragraph': sharedInnerBlocks['core/paragraph'],
-		'elements/link': sharedInnerBlocks['elements/link'],
-		'core/heading': sharedInnerBlocks['core/heading'],
+		'core/paragraph': {
+			...sharedInnerBlocks['core/paragraph'],
+			settings: {
+				...sharedInnerBlocks['core/paragraph'].settings,
+				force: true,
+			},
+		},
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			settings: {
+				...sharedInnerBlocks['elements/link'].settings,
+				force: false,
+			},
+		},
+		'core/heading': {
+			...sharedInnerBlocks['core/heading'],
+			settings: {
+				...sharedInnerBlocks['core/heading'].settings,
+				force: false,
+			},
+		},
 		'core/heading-1': sharedInnerBlocks['core/heading-1'],
 		'core/heading-2': sharedInnerBlocks['core/heading-2'],
 		'core/heading-3': sharedInnerBlocks['core/heading-3'],